add flags to initstatics.go
