+++ pkg/test/framework/initstatics.go
@@ -20,4 +20,5 @@ import (
 	_ "istio.io/istio/pkg/test/framework/components/echo"
 	_ "istio.io/istio/pkg/test/framework/components/environment/kube"
 	_ "istio.io/istio/pkg/test/framework/components/istio"
+	_ "istio.io/istio/pkg/test/framework/components/telemetry"
 )
