+++ tests/integration/telemetry/stats/prometheus/nullvm/istioctl_metrics_test.go
@@ -23,8 +23,8 @@ import (
 
 	"istio.io/istio/pkg/test/framework"
 	"istio.io/istio/pkg/test/framework/components/istioctl"
+	telemetrypkg "istio.io/istio/pkg/test/framework/components/telemetry"
 	"istio.io/istio/pkg/test/util/retry"
-	"istio.io/istio/tests/integration/telemetry"
 	common "istio.io/istio/tests/integration/telemetry/stats/prometheus"
 )
 
@@ -41,7 +41,7 @@ func TestIstioctlMetrics(t *testing.T) {
 					return err
 				}
 				return validateDefaultOutput(t, ctx, "server")
-			}, retry.Delay(telemetry.RetryDelay), retry.Timeout(telemetry.RetryTimeout))
+			}, retry.Delay(telemetrypkg.RetryDelay), retry.Timeout(telemetrypkg.RetryTimeout))
 
 		})
