+++ tests/integration/telemetry/stats/prometheus/stats.go
@@ -30,6 +30,7 @@ import (
 	"istio.io/istio/pkg/test/framework/components/istio/ingress"
 	"istio.io/istio/pkg/test/framework/components/namespace"
 	"istio.io/istio/pkg/test/framework/components/prometheus"
+	"istio.io/istio/pkg/test/framework/components/telemetry"
 	"istio.io/istio/pkg/test/framework/features"
 	"istio.io/istio/pkg/test/framework/resource"
 	"istio.io/istio/pkg/test/util/retry"
@@ -104,7 +105,7 @@ func TestStatsFilter(t *testing.T, feature features.Feature) {
 						}
 
 						return nil
-					}, retry.Delay(util.RetryDelay), retry.Timeout(util.RetryTimeout))
+					}, retry.Delay(telemetry.RetryDelay), retry.Timeout(telemetry.RetryTimeout))
 					if err != nil {
 						return err
 					}
@@ -139,7 +140,7 @@ func TestStatsTCPFilter(t *testing.T, feature features.Feature) {
 						}
 
 						return nil
-					}, retry.Delay(util.RetryDelay), retry.Timeout(util.RetryTimeout))
+					}, retry.Delay(telemetry.RetryDelay), retry.Timeout(telemetry.RetryTimeout))
 					if err != nil {
 						return err
 					}
